pyproject.toml
[project]
name = "industria_moveleira_backend"
version = "0"
description = "Projeto Integrador"
authors = [
    {name = "doginvi", email = "victordesousaestudo@gmail.com"},
]
dependencies = ["django>=5.2.3", "djangorestframework>=3.16.0"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
